mptcp: keep the family test in if and else the same
In __mptcp_subflow_connect(), the address family is checked to be AF_INET
or AF_INET6. But the family tested in 'if' and 'else' are inconsistent.
In 'if', local->addr.family is tested, while in 'else', sk->sk_family is
tested. This is a bit strange:

	if (local->addr.family == AF_INET) {
		...
	} else if (sk->sk_family == AF_INET6) {
		...
	}

Although the results are the same, it's more natural to keep the family of
the test in 'if' and 'else' the same:

	if (local->addr.family == AF_INET) {
		...
	} else if (local->addr.family == AF_INET6) {
		...
	}

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>