mptcp: pm: drop redundant mptcp_pm_send_ack
mptcp_pm_send_ack() is called twice in __mptcp_pm_addr_send_ack(), which
makes the code a bit redundant. The first call to mptcp_pm_send_ack() when
a non-stale subflow is found in the loop can be removed. Instead, we can
break the loop to use the second mptcp_pm_send_ack().

Signed-off-by: Geliang Tang <tanggeliang@kylinos.cn>